+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 ### Changed
   - Rimossa redirect in caso di cancellazione di un orario di lavoro particolare che generava loop infinito
+  - Resi opzionali parametri type e year nel ical dei turni
 
 ## [2.12.0] - 2024-01-15
 ### Added